feat: update screens and file structure (#99)
