+++ components/ArchivedGamesContainer/ArchivedGamesContainer.jsx
@@ -1,7 +1,7 @@
 import { View, Text, StyleSheet, ScrollView } from "react-native";
 import React, { useState, useEffect } from "react";
-import ArchivedGameCard from "./ArchivedGameCard";
 import axiosInstance from "@/services/axios/axiosInstance";
+import ArchivedGameCard from "@/components/ArchivedGamesContainer/ArchivedGameCard";
 
 const ArchivedGamesContainer = ({
   navigation,
@@ -38,8 +38,6 @@ const ArchivedGamesContainer = ({
 
   return (
     <View style={styles.archivedContainer}>
-      <GamesHeader />
-
       <ScrollView>
         {archivedGames.map((game, index) => {
           return (
@@ -57,44 +55,10 @@ const ArchivedGamesContainer = ({
   );
 };
 
-const GamesHeader = () => {
-  return (
-    <View style={styles.gameHeader}>
-      <Text style={styles.gameHeaderText}>Opponent</Text>
-      <View style={styles.resultAndMoves}>
-        <Text style={styles.gameHeaderText}>Result</Text>
-        <Text style={styles.gameHeaderText}>Moves</Text>
-      </View>
-    </View>
-  );
-};
-
 const styles = StyleSheet.create({
   archivedContainer: {
     width: "100%",
-    height: "60%",
-    flexDirection: "column",
-    marginBottom: 20
-  },
-  gameHeader: {
-    width: "95%",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginBottom: 3,
-    marginLeft: 10
-  },
-  gameHeaderText: {
-    fontSize: 20,
-    color: "white"
-  },
-  dateAndSort: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 5
-  },
-  resultAndMoves: {
-    flexDirection: "row",
-    gap: 15
+    flexDirection: "column"
   }
 });
