+++ components/RightSideBar/ReviewGameBarStyles.js
@@ -6,7 +6,7 @@ const styles = StyleSheet.create({
     height: "100%",
     width: "100%",
     justifyContent: "space-between",
-    backgroundColor: "black"
+    backgroundColor: "#252525"
   }
 });
