+++ screens/CategoryStats/CategoryStatsModalStyling.js
@@ -34,7 +34,7 @@ const styles = StyleSheet.create({
     flexDirection: "row",
     justifyContent: "space-between",
     alignItems: "center",
-    width: "100%",
+    width: "100%"
   },
   categoryText: {
     fontSize: 25,
