+++ screens/CategoryStats/GameStat.jsx
@@ -1,13 +1,17 @@
 import { useRoute } from "@react-navigation/native";
 import { View, Text, Pressable } from "react-native";
 import React, { useState, useEffect } from "react";
-import ArchivedGamesContainer from "../ProfileScreen/ArchivedGamesContainer";
 import styles from "./gameStatStyles";
 import * as SecureStore from "expo-secure-store";
 import axiosInstance from "@/services/axios/axiosInstance";
 import Header from "@/components/Header/Header";
 import Icon from "react-native-vector-icons/FontAwesome";
 import CategoryStatsModal from "@/screens/CategoryStats/CategoryStatsModal";
+import { SafeAreaView } from "react-native-safe-area-context";
+import TimeCategoryRecordCards from "@/components/TimeCategoryRecordCards/TimeCategoryRecordCards";
+import FontAwesome from "@expo/vector-icons/FontAwesome";
+import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
+import ArchivedGamesContainer from "@/components/ArchivedGamesContainer/ArchivedGamesContainer";
 
 const GameStat = ({ navigation }) => {
   const route = useRoute();
@@ -44,31 +48,51 @@ const GameStat = ({ navigation }) => {
     setIsModalOpen(true);
   };
 
+  const getIcon = () => {
+    if (category === "bullet") {
+      return <MaterialCommunityIcons name="bullet" size={25} color="yellow" />;
+    } else if (category === "blitz") {
+      return <FontAwesome name="bolt" size={25} color="yellow" />;
+    }
+
+    return <MaterialCommunityIcons name="timer-outline" size={25} color="green" />
+  }
+
   return (
-    <View style={styles.GameStat}>
+    <SafeAreaView style={styles.GameStat}>
       <Header navigation={navigation} />
-      <View style={styles.headerContainer}>
+      <Pressable style={styles.headerContainer} onPress={onPress}>
+        {getIcon()}
         <Text style={styles.timeControlTitle}>{category.toUpperCase()}</Text>
-        <Pressable onPress={onPress}>
-          <Icon name="chevron-down" size={15} color="white" />
-        </Pressable>
-      </View>
+        <Icon name="chevron-down" size={15} color="white" />
+      </Pressable>
 
       <Text style={styles.usernameHeader}>{username}</Text>
       {gameStats ? (
         <View style={styles.diagnostics}>
-          <Text style={styles.diagnosticsText}>Wins: {gameStats.wins}</Text>
-          <Text style={styles.diagnosticsText}>Losses: {gameStats.losses}</Text>
-          <Text style={styles.diagnosticsText}>Draws: {gameStats.draws}</Text>
-          <Text style={styles.diagnosticsText}>Rating: {gameStats.rating}</Text>
+          <View style={styles.ratingContainer}>
+            <Text style={styles.statLabel}>Rating</Text>
+            <Text style={[styles.diagnosticsText, styles.ratingText]}>
+              {gameStats.rating}
+            </Text>
+          </View>
+
+          <TimeCategoryRecordCards
+            wins={gameStats.wins}
+            draws={gameStats.draws}
+            losses={gameStats.losses}
+          />
         </View>
       ) : (
         <Text>Loading...</Text>
       )}
+
       <Text style={styles.archivedGamesHeader}>Games Played</Text>
+
       <ArchivedGamesContainer
         playerUsername={username}
         timeControl={timeControl}
+        navigation={navigation}
       />
 
       {isModalOpen && (
@@ -78,7 +102,7 @@ const GameStat = ({ navigation }) => {
           setCategory={setCategory}
         />
       )}
-    </View>
+    </SafeAreaView>
   );
 };
