+++ screens/HomeScreen/playGameOptionsStyles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    cardContainer: {
+        flexDirection: "row",
+        justifyContent: "center",
+        alignItems: "center",
+        flexWrap: "wrap"
+    }
+});
+
+export default styles;
