+++ screens/RegisterScreen/Register.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from "react";
-import { View, Text, TextInput, Button, StyleSheet } from "react-native";
+import { View, Text, TextInput, Button } from "react-native";
 import axiosInstance from "@/services/axios/axiosInstance";
 import { SafeAreaView } from "react-native-safe-area-context";
+import styles from "./registerStyling"
 
 const Register = ({ navigation }) => {
   const [email, setEmail] = useState("");
@@ -81,47 +82,5 @@ const Register = ({ navigation }) => {
     </SafeAreaView>
   );
 };
-
-const styles = StyleSheet.create({
-  registerHeader: {
-    fontSize: 30,
-    fontWeight: "bold",
-    marginBottom: 10,
-    color: "white"
-  },
-  registerContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#252525"
-  },
-  registerCard: {
-    borderRadius: 10,
-    overflow: "hidden",
-    width: "60%"
-  },
-  gradient: {
-    flex: 1,
-    borderRadius: 10,
-    borderWidth: 2,
-    borderColor: "transparent"
-  },
-  innerCard: {
-    backgroundColor: "#121212",
-    padding: 20
-  },
-  input: {
-    backgroundColor: "white",
-    marginBottom: 10,
-    padding: 10,
-    borderRadius: 5
-  },
-  errorCard: {
-    color: "white",
-    fontSize: 20,
-    marginTop: 10,
-    borderRadius: 10
-  }
-});
 
 export default Register;
